[ci] fix #294: diameter truncated to int on weighted graphs
Graph::m_graphDiameter (and graphDiameter()/graphDiameterCached()/
graphDiameterSigned()) was int throughout, so on a weighted graph the true
fractional diameter silently lost its fractional part. avg_distance, CC,
eccentricity and distance_sum already reported the exact weighted value -
diameter was the one inconsistent metric, and the GUI's own dialog text
("the diameter can be greater than N" for weighted networks) already implies
the weighted metric was the intended one.

Changed to qreal end to end: the Graph field/API, the engine's per-source
max-distance tracking (ThreadLocalState::maxDiameter, sourceMaxDist), the
CLI distance kernel's JSON/compare logic (now tolerance-compared like
avg_distance instead of exact-int), and the two GUI call sites. Diameter now
tracks the same metric as every other distance-derived value: hop count when
unweighted, the true weighted geodesic distance when considerWeights=1 -
matching networkx/igraph/UCINET convention.

Independently verified the new fractional values (DunbarGelada weighted
diameter = 1.1011904761904763) against shortest-path distances computed
outside SocNetV. 5 existing baselines re-dumped.

+++ src/engine/distance_engine.cpp
@@ -755,7 +755,7 @@ void DistanceEngine::runAllSources(const bool computeCentralities,
         // ones) - tls.pss.dist[vi] already holds RAND_MAX for every unreached vi
         // (PerSourceScratch::resetPerSource() fills it before every source, unconditionally),
         // so this isn't new work - it reuses a reset that was already happening.
-        int sourceMaxDist = 0;
+        qreal sourceMaxDist = 0;
         qreal sourceDistanceSum = 0;
         for (int vi = 0; vi < totalV; ++vi)
         {
@@ -766,7 +766,7 @@ void DistanceEngine::runAllSources(const bool computeCentralities,
             {
                 sourceDistanceSum += tls.pss.dist[vi];
                 if (tls.pss.dist[vi] > sourceMaxDist)
-                    sourceMaxDist = (int)tls.pss.dist[vi];
+                    sourceMaxDist = tls.pss.dist[vi];
             }
         }
         if (sourceMaxDist > tls.maxDiameter)
